Refactor navigation components and simplify component structure
