+++ FABLECRAFT NUXT BUILD/components/atoms/NavigationItem.vue
